修改更新

+++ flink-test-utils-parent/flink-test-utils/src/main/java/org/apache/flink/test/util/TestBaseUtils.java
@@ -18,17 +18,21 @@
 
 package org.apache.flink.test.util;
 
+import org.apache.flink.api.common.time.Time;
 import org.apache.flink.api.java.tuple.Tuple;
 import org.apache.flink.configuration.AkkaOptions;
 import org.apache.flink.configuration.ConfigConstants;
 import org.apache.flink.configuration.Configuration;
+import org.apache.flink.configuration.CoreOptions;
 import org.apache.flink.configuration.HighAvailabilityOptions;
-import org.apache.flink.configuration.JobManagerOptions;
 import org.apache.flink.configuration.TaskManagerOptions;
+import org.apache.flink.configuration.WebOptions;
 import org.apache.flink.core.testutils.CommonTestUtils;
 import org.apache.flink.runtime.messages.TaskManagerMessages;
 import org.apache.flink.runtime.minicluster.LocalFlinkMiniCluster;
 import org.apache.flink.util.TestLogger;
+
+import org.apache.flink.shaded.netty4.io.netty.handler.codec.http.HttpResponseStatus;
 
 import akka.actor.ActorRef;
 import akka.dispatch.Futures;
@@ -36,10 +40,11 @@ import akka.pattern.Patterns;
 import akka.util.Timeout;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
-import org.apache.hadoop.fs.FileSystem;
 import org.junit.Assert;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import javax.annotation.Nonnull;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedReader;
@@ -49,6 +54,7 @@ import java.io.FileReader;
 import java.io.FilenameFilter;
 import java.io.IOException;
 import java.io.InputStream;
+import java.lang.reflect.Method;
 import java.net.HttpURLConnection;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -63,7 +69,9 @@ import java.util.Comparator;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -87,13 +95,18 @@ public class TestBaseUtils extends TestLogger {
 
 	protected static final int MINIMUM_HEAP_SIZE_MB = 192;
 
-	protected static final long TASK_MANAGER_MEMORY_SIZE = 80;
+	protected static final String TASK_MANAGER_MEMORY_SIZE = "80m";
 
 	protected static final long DEFAULT_AKKA_ASK_TIMEOUT = 1000;
 
 	protected static final String DEFAULT_AKKA_STARTUP_TIMEOUT = "60 s";
 
 	public static final FiniteDuration DEFAULT_TIMEOUT = new FiniteDuration(DEFAULT_AKKA_ASK_TIMEOUT, TimeUnit.SECONDS);
+
+	public static final Time DEFAULT_HTTP_TIMEOUT = Time.seconds(10L);
+
+	static final String NEW_CODEBASE = "new";
+	static final String CODEBASE_KEY = "codebase";
 
 	// ------------------------------------------------------------------------
 
@@ -119,7 +132,7 @@ public class TestBaseUtils extends TestLogger {
 		Configuration config = new Configuration();
 
 		config.setInteger(ConfigConstants.LOCAL_NUMBER_TASK_MANAGER, numTaskManagers);
-		config.setInteger(ConfigConstants.TASK_MANAGER_NUM_TASK_SLOTS, taskManagerNumSlots);
+		config.setInteger(TaskManagerOptions.NUM_TASK_SLOTS, taskManagerNumSlots);
 
 		config.setBoolean(ConfigConstants.LOCAL_START_WEBSERVER, startWebserver);
 
@@ -135,22 +148,41 @@ public class TestBaseUtils extends TestLogger {
 		Configuration config,
 		boolean singleActorSystem) throws Exception {
 
-		logDir = File.createTempFile("TestBaseUtils-logdir", null);
-		Assert.assertTrue("Unable to delete temp file", logDir.delete());
-		Assert.assertTrue("Unable to create temp directory", logDir.mkdir());
-		Path logFile = Files.createFile(new File(logDir, "jobmanager.log").toPath());
-		Files.createFile(new File(logDir, "jobmanager.out").toPath());
+		if (!config.contains(WebOptions.LOG_PATH) || !config.containsKey(ConfigConstants.TASK_MANAGER_LOG_PATH_KEY)) {
+			logDir = File.createTempFile("TestBaseUtils-logdir", null);
+			Assert.assertTrue("Unable to delete temp file", logDir.delete());
+			Assert.assertTrue("Unable to create temp directory", logDir.mkdir());
+			Path logFile = Files.createFile(new File(logDir, "jobmanager.log").toPath());
+			Files.createFile(new File(logDir, "jobmanager.out").toPath());
 
-		config.setLong(TaskManagerOptions.MANAGED_MEMORY_SIZE, TASK_MANAGER_MEMORY_SIZE);
-		config.setBoolean(ConfigConstants.FILESYSTEM_DEFAULT_OVERWRITE_KEY, true);
+			if (!config.contains(WebOptions.LOG_PATH)) {
+				config.setString(WebOptions.LOG_PATH, logFile.toString());
+			}
 
-		config.setString(AkkaOptions.ASK_TIMEOUT, DEFAULT_AKKA_ASK_TIMEOUT + "s");
-		config.setString(AkkaOptions.STARTUP_TIMEOUT, DEFAULT_AKKA_STARTUP_TIMEOUT);
+			if (!config.containsKey(ConfigConstants.TASK_MANAGER_LOG_PATH_KEY)) {
+				config.setString(ConfigConstants.TASK_MANAGER_LOG_PATH_KEY, logFile.toString());
+			}
+		}
 
-		config.setInteger(JobManagerOptions.WEB_PORT, 8081);
-		config.setString(JobManagerOptions.WEB_LOG_PATH, logFile.toString());
+		if (!config.contains(WebOptions.PORT)) {
+			config.setInteger(WebOptions.PORT, 8081);
+		}
 
-		config.setString(ConfigConstants.TASK_MANAGER_LOG_PATH_KEY, logFile.toString());
+		if (!config.contains(AkkaOptions.ASK_TIMEOUT)) {
+			config.setString(AkkaOptions.ASK_TIMEOUT, DEFAULT_AKKA_ASK_TIMEOUT + "s");
+		}
+
+		if (!config.contains(AkkaOptions.STARTUP_TIMEOUT)) {
+			config.setString(AkkaOptions.STARTUP_TIMEOUT, DEFAULT_AKKA_STARTUP_TIMEOUT);
+		}
+
+		if (!config.contains(CoreOptions.FILESYTEM_DEFAULT_OVERRIDE)) {
+			config.setBoolean(CoreOptions.FILESYTEM_DEFAULT_OVERRIDE, true);
+		}
+
+		if (!config.contains(TaskManagerOptions.MANAGED_MEMORY_SIZE)) {
+			config.setString(TaskManagerOptions.MANAGED_MEMORY_SIZE, TASK_MANAGER_MEMORY_SIZE);
+		}
 
 		LocalFlinkMiniCluster cluster =  new LocalFlinkMiniCluster(config, singleActorSystem);
 
@@ -204,7 +236,17 @@ public class TestBaseUtils extends TestLogger {
 			}
 
 			executor.stop();
-			FileSystem.closeAll();
+			try {
+				Class<?> hadoopFileSystemClass = Class.forName(
+					"org.apache.hadoop.fs.FileSystem",
+					true,
+					TestBaseUtils.class.getClassLoader());
+
+				Method closeAllMethod = hadoopFileSystemClass.getMethod("closeAll");
+				closeAllMethod.invoke(null);
+			} catch (Throwable e) {
+				// ignore
+			}
 			System.gc();
 
 			Assert.assertEquals("Not all broadcast variables were released.", 0, numUnreleasedBCVars);
@@ -442,7 +484,7 @@ public class TestBaseUtils extends TestLogger {
 				throw new IllegalArgumentException("This path does not denote a local file.");
 			}
 		} catch (URISyntaxException | NullPointerException e) {
-			throw new IllegalArgumentException("This path does not describe a valid local file URI.");
+			throw new IllegalArgumentException("This path does not describe a valid local file URI.", e);
 		}
 	}
 
@@ -600,21 +642,60 @@ public class TestBaseUtils extends TestLogger {
 	//---------------------------------------------------------------------------------------------
 
 	public static String getFromHTTP(String url) throws Exception {
-		URL u = new URL(url);
+		return getFromHTTP(url, DEFAULT_HTTP_TIMEOUT);
+	}
+
+	public static String getFromHTTP(String url, Time timeout) throws Exception {
+		final URL u = new URL(url);
 		LOG.info("Accessing URL " + url + " as URL: " + u);
-		HttpURLConnection connection = (HttpURLConnection) u.openConnection();
-		connection.setConnectTimeout(100000);
-		connection.connect();
-		InputStream is;
-		if (connection.getResponseCode() >= 400) {
-			// error!
-			LOG.warn("HTTP Response code when connecting to {} was {}", url, connection.getResponseCode());
-			is = connection.getErrorStream();
-		} else {
-			is = connection.getInputStream();
+
+		final long deadline = timeout.toMilliseconds() + System.currentTimeMillis();
+
+		while (System.currentTimeMillis() <= deadline) {
+			HttpURLConnection connection = (HttpURLConnection) u.openConnection();
+			connection.setConnectTimeout(100000);
+			connection.connect();
+
+			if (Objects.equals(HttpResponseStatus.SERVICE_UNAVAILABLE, HttpResponseStatus.valueOf(connection.getResponseCode()))) {
+				// service not available --> Sleep and retry
+				LOG.debug("Web service currently not available. Retrying the request in a bit.");
+				Thread.sleep(100L);
+			} else {
+				InputStream is;
+
+				if (connection.getResponseCode() >= 400) {
+					// error!
+					LOG.warn("HTTP Response code when connecting to {} was {}", url, connection.getResponseCode());
+					is = connection.getErrorStream();
+				} else {
+					is = connection.getInputStream();
+				}
+
+				return IOUtils.toString(is, ConfigConstants.DEFAULT_CHARSET);
+			}
 		}
 
-		return IOUtils.toString(is, connection.getContentEncoding() != null ? connection.getContentEncoding() : "UTF-8");
+		throw new TimeoutException("Could not get HTTP response in time since the service is still unavailable.");
+	}
+
+	@Nonnull
+	public static CodebaseType getCodebaseType() {
+		return Objects.equals(NEW_CODEBASE, System.getProperty(CODEBASE_KEY)) ? CodebaseType.NEW : CodebaseType.LEGACY;
+	}
+
+	public static boolean isNewCodebase() {
+		return CodebaseType.NEW == getCodebaseType();
+	}
+
+	/**
+	 * Type of the mini cluster to start.
+	 *
+	 * @deprecated Will be irrelevant once the legacy mode has been removed.
+	 */
+	@Deprecated
+	public enum CodebaseType {
+		LEGACY,
+		NEW
 	}
 
 	/**
